requirements.txt
django==1.11.17
django-filter==1.0.4
djangorestframework==3.6.3
gunicorn==19.7.1
psycopg2-binary
django-storages==1.6.3
boto==2.48.0
slackclient==1.0.6
sentry-sdk==0.6.1
python-dateutil
appointmentguru_helpers

kong_oauth==0.1

requests
responses

ipdb
ipython
Faker
coverage
codacy-coverage
flask
